minicoredumper: add compression feature
A new recept option "compression" is available. If configured, the
core file will be placed within a sparse tar file and compressed
using the configured compressor. When this option is configured,
the vast majority of the core file and the entire compressed tar
archive are created in memory and via stdio pipes. Therefore, this
feature is particularly interesting for systems with very little
dump space or dump space that does not support sparse files.

Signed-off-by: John Ogness <john.ogness@linutronix.de>