Format timestamp as YYYYMMDD.HHMMSS+ZZZZ
I am not sure what format is the best, but I think that having a human
readable timestamp should greatly help in identifying which dump is
relevant.  I am also not sure if a format with the timezone should
be used, or if a UTC time should always be used.

The resulting directory looks like:

  a.out.20151005.113619-0400.9811

Also, if either of localtime and strftime fails for some reason, we could
fall back on a simpler name, since have a "badly" named core dump would be
more useful than having no core dump at all.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>