dump_vma: use mem_end instead of file_end
The actual end of the section in memory is saved in mem_end.  This causes
some important information that GDB could use to be lost.  Also, it makes
sense to use mem_end there if it is used in get_vma_pos.

Additionally, it might be a good idea to add a log in:

	/* make sure we have something to dump */
	if (start >= end)
		return 0;

as this is probably a situation that should not happen, and if it happens
it might be a bug.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>