Dump NUL character at the end of solib name
When dumping the path of a shared object, minicoredumper dumps all the
characters of the string, but not the terminating NUL character.  When
we implement support for unavailable bytes in GDB, it is suddenly unable
to read the path from the core.  This is because it now trips on an
unavailable byte, where before it would read it as a 0.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>