Fix g_type_init warning
When compiling with glib >= 2.36, we get this warning:

main.c:115:2: warning: ‘g_type_init’ is deprecated (declared at /usr/include/glib-2.0/gobject/gtype.h:667) [-Wdeprecated-declarations]
  g_type_init();

It is necessary to call it for earlier versions though.  The strategy
used by other projects I contributed to was to make "compatibility"
functions to deal with different version of libs or the kernel.  This
has the advantage of not polluting the code with #ifdefs.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>