minicoredumper: add elf section note with dump list
A section note ".note.minicoredumper.dumplist" is added to the
core file. The note contains a list of offsets and lengths of
dump areas included in the core file. This can be used by tools
such as gdb in order to distinguish between data that has a
value of 0 and data that was not dumped.

NOTE: This feature relies on elfutils release 0.167 or higher. In
      particular, commit 96e140f6687922606657a76f185a73cf47908ef2
      of the elfutils git repository.

Signed-off-by: John Ogness <john.ogness@linutronix.de>