minicoredumper: use shared memory object for ELF work
Since 80f8e1d681609172f806e57d39dc40fdadc94f87 it is expected that
the base_dir is located on a filesystem that supports sparse files.
That may not be the case.

Rather than using regular files, the ELF parsing/extending is now
done on a shared memory object. This should guarantee the
sparseness required for that feature.

Signed-off-by: John Ogness <john.ogness@linutronix.de>