libminicoredumper: fix inotify
A return value check appeared in 878e2bb4e8bbfe26d1c8d18df94dbf5dba252ac1
but was incorrect, thus breaking inotify. Since it is not known in
advance which file is involved with the next event, the read() is done
purposely with a buffer that is larger than what read() will actually
return. It should, however, be at least of size struct inotify_event.

Signed-off-by: John Ogness <john.ogness@linutronix.de>