api, (lib)minicoredumper: change api to match scanf syntax
The variable arguments of mcd_dump_data_register_text() nearly
match the syntax of scanf(), with the exception of strings. This
can be confusing to the developer. The API has been changed to
match scanf() for strings as well.

Since this change also modifies how strings are stored internally
in memory, DUMP_DATA_VERSION is bumped to 2.

Signed-off-by: John Ogness <john.ogness@linutronix.de>