libminicoredumper: fix size calculation
The size value for each registered text variable was always set
to the size of a pointer instead of the size of the item pointed
to. Although a bug, it went unnoticed because strings were also
stored as a pointer to the string pointer.

Since 4c0f76639ab26b3830424ac67732769a44a52124 strings are now
stored as a pointer to the first string character, thus this bug
is clearly visible.

Now sizes are stored based on their type and strings based on
their length (+1).

Signed-off-by: John Ogness <john.ogness@linutronix.de>