minicoredumper: reverse symbol source list
If libraries and the executable use a symbol, the symbol location
from the executable will be used. Since minicoredumper uses the
first symbol location it finds, the list of symbol sources has been
adjusted so that the executable symbols are checked before the library
symbols.

It is assumed that if this situation occurs between libraries,
the library with the lower address is the one to look at. This
assumption should be investigated further.

Signed-off-by: John Ogness <john.ogness@linutronix.de>