autoconf: use autotools for all configuration and installation
Through the use of autoconf it is now possible to configure
various values such as dump path, configuration path, regd user.
The "install" target will perform a full installation (using a
SysV script). The default configuration is set for a minimal yet
useful out-of-the-box experience. With these changes, a user can:

    ./configure && make && sudo make install

and have a useable minicoredumper ready to go.

Signed-off-by: John Ogness <john.ogness@linutronix.de>