debian: move debianization into a dists directory
Distribution packaging is really a job for the package maintainer.
However, we would like to keep a copy ourselves in git. So a new
"dists" sub-directory is now available for that.

Also, since autotools now does nearly all installation work and
uses decent defaults, there is no reason for the Debian packaging
to do it. This simplifies the Debian packaging.

Signed-off-by: John Ogness <john.ogness@linutronix.de>