autoconf/debian: cleanup manpage and script handling
Manpages and scripts were being specified directly into dist_* variables.
That is now how such files should be specified. It also complicated the
process of packaging minicoredumper for Debian.

Now manpages and scripts are specified in man_MANS and sbin_SCRIPTS.
The Debian package now makes use of the symlink feature of dh_installman
to handle the symbolic link manpage.

Signed-off-by: John Ogness <john.ogness@linutronix.de>