minicoredumper: explicitly limit checksum value
With -Werror=format-truncation, gcc reports that possible snprintf
trucation occurs for the checksum value. If gcc was a little more
intelligent, it would see that such a truncation can never happen.

Add an extra check (that can never return true) to make it clear
to gcc that no value can be returned that will cause truncation.

Reported-by: debian-gcc@lists.debian.org
Reported-by: Stefan Bosak <stefan.bosak@gmail.com>
Signed-off-by: John Ogness <john.ogness@linutronix.de>
Closes: Debian Bug#897810