minicoredumper: retry elf parsing as long as needed
As was reported in github issue #2 ("maximum number of tries
insufficient, in rare cases, for elf parse"), the number of retries
for parsing a process may be insufficient. Rather than setting an
upper limit on the maximum number of retries, track the number of
headers seen. As long as the number of seen headers is greater than
the previous try, try again.

In order to avoid introducing any new issues, preserve the behavior
of retrying at least 10 times, even if no new headers are seen.

Reported-by: github.com/ssajal-wr
Signed-off-by: John Ogness <john.ogness@linutronix.de>