minicoredumper: handle all read_remote() failures
Due to memory corruption or unexpected layouts, it is possible that
read_remote() could fail. For this reason it is important that
read_remote() return values are always checked.

In particular, if read_remote() fails within init_from_auxv(),
the minicoredumper could enter into an infinite loop. This not only
hogs a CPU, but will also flood syslog.

Signed-off-by: John Ogness <john.ogness@linutronix.de>