minicoredumper: do not dump beyond vma
For some of the memory mapped regions, the virtual address range
is larger than the core file size. The minicoredumper has a check
to prevent dumping beyond a memory mapped region, but this check
was based on the virtual address range rather than the core file
range. This led to core file corruption.

Change the check to be based on the core file size.

This problem is only seen when such maps are explicitly selected
within the recept file. By default, minicoredumper does not dump
such maps.

Reported-by: Abhi Arjun <abhiarjun@tutanota.com>
Signed-off-by: John Ogness <john.ogness@linutronix.de>