init: set max value in kernel.core_pipe_limit
The kernel defaults to 0 for kernel.core_pipe_limit. This special
value works because the kernel _does_ wait as long as the core file
pipe is open, which minicoredumper does not close until it is
finished.

However, this may be a kernel bug and it should not be relied upon.
Set kernel.core_pipe_limit to the maximum value on start and restore
it to 0 on stop.

Signed-off-by: John Ogness <john.ogness@linutronix.de>