minicoredumper: dump zeros in 4kb blocks
Dumping zeros 1 byte at a time is horribly inefficient. If tar
files are used, this does not matter much. But it tar files are
not used, there may be gigabytes of zeros that need to be dumped.

Dumping the zeros in 4kb blocks is a huge improvement. Using
larger buffer sizes does not seem to add any noticeable speed
benefits.

Signed-off-by: John Ogness <john.ogness@linutronix.de>