minicoredumper: truncate tar core files to 8GB
The ustar format only has 11 octal characters available for specifying
sizes and offsets. So the maximum value is 077777777777 (8589934591).

Drop core data blocks and truncate the core file if it goes beyond this
limit. This is only a workaround until a more modern tar format is
implemented.

The discussion about this issue and the possible workarounds can be
seen on the mailing list:

https://lists.linutronix.de/pipermail/minicoredumper/2023-February/000126.html

In that thread it is also mentioned how an application can guarantee
thread stacks will not be truncated by placing them below the
anticipated heap area, thus placing them near the beginning of the
core file.

Reported-by: github.com/sehdshskj
Reported-by: Jos Hulzink (Ellips B.V.) <jos.hulzink@ellips.com>
Signed-off-by: John Ogness <john.ogness@linutronix.de>