minicoredumper: add eof check for read_file_fd()
Commit a09461321cb8 ("minicoredumper: treat eof as error") added an eof
check to the higher-level copy_data() function, assuming that the
lower-level read_file_fd() could return 0. But actually that function
was not handling the eof and would instead go into an infinite loop.

read() can return 0 if the crashing process is killed (for example with
explicit SIGKILL) before the minicoredumper is finished reading the
memory.

The problem was reported and fixed by Dongyang. The original report:

    In the below situation, the read will in an infinite loop:

    If the PID 123 triggered the core dump and Minicoredumper started
    to read the /proc/123/mem (dump_compressed_tar-->copy_data-->read_file_fd).
    At this time, if the OOM is triggered, then the PID 123 is killed by
    SIGKILL. The /proc/123/ folder will nonexistent, the read will return zero. 

    But in the read_file_fd, the code doesn't check the zero condition.
    When the "r" is 0,  the "len" will never to 0,  then into an infinite loop. 

Fixes: a09461321cb8 ("minicoredumper: treat eof as error")
Signed-off-by: Dongyang <dongyang626@gmail.com>
Signed-off-by: Jianyu <jianyu9785@gmail.com>
Signed-off-by: John Ogness <john.ogness@linutronix.de>