minicoredumper: support comm values with spaces
A comm string may contain spaces, which would cause the
minicoredumper to choose the wrong item for the thread
stack when parsing proc/stat. This results in the
minicoredumper failing to copy the stack for some threads.

For the comm item, explicitly check for the terminating
parenthesis before continuing to the next item. Note
that the error is still possible if ") " is within
comm. But this is really a limitation of the primitive
format of proc/stat.

Signed-off-by: John Ogness <john.ogness@linutronix.de>