minicoredumper: fix dump of robust mutex list with pi
glibc uses bit0 of the robust mutex list pointers as a flag for
PI (priority inheritance) mutexes. The minicoredumper was not
masking the flag before trying to interpret the address. This
led to the wrong pointer data being dumped into the core file
if any of the locked robust mutexese are using the
PTHREAD_PRIO_INHERIT attribute.

Mask bit0 before interpretting the robust mutex list pointers.

Also cleanup the list iterating so that it is easier to follow
and does not treat the head as an extra mutex.

Signed-off-by: John Ogness <john.ogness@linutronix.de>