minicoredumper: only dump to core for crash process
dump_data_content() is used to dump core and external data
files for both the crashing process and any other applications
registered with the minicoredumper.

For the non-crashing process there is no core file to dump to,
but dump_data_content() does not consider this and attempts
anyway. This leads to error messages that can mislead the user
to think there was a problem.

Adjust dump_data_content() to only attempt dumping to a core
file if a core file is available.

Signed-off-by: John Ogness <john.ogness@linutronix.de>